+++ src/components/Image/Image.stories.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import viewsPalette from 'theme/palette';
+import palette from 'theme/palette';
 import Image from './Image';
 
 const control = {
 	type: 'select',
-	options: Object.keys(viewsPalette).reduce((options, colorName) => {
-		options[colorName] = viewsPalette[colorName];
+	options: Object.keys(palette).reduce((options, colorName) => {
+		options[colorName] = palette[colorName];
 		return options;
 	}, {})
 };