+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- export palette colors
+
+### Changed
+
+- viewPalette to palette name
+
 ## [0.20.0] - 2024-08-27
 
 ### Added